first logger

+++ logger/src/ds18b20.rs
@@ -10,8 +10,8 @@ static W1_PATH_SUFFIX: &str = "w1_slave";
 pub struct MilliCelsius(u32);
 impl MilliCelsius {
 
-    pub fn to_celsius(self) -> f64 {
-        (self.0 as f64) / 1000.0 
+    pub fn to_celsius(self) -> f32 {
+        (self.0 as f32) / 1000.0 
     }
 }
